Server full fix + refactor + moving ProxyRepository and Serice to SharedClassLibrary (alex + endre + marius modifications) (#20)
* Read BaseApiUrl from appsettings.json

You should also add:
```
{
  "BaseApiUrl": "https://localhost:7194"
}
```
, or the server address to MarketPlace924/appsettings.json

* Copy appsettings.json to the output directory on build

This is needed because the app is looking for appsettings.json in the output directory (the same directory as the executable).

* Implement UserRepository proxy and controller

* Update Server/Controllers/UserApiController.cs


* delete appsettings

* Update .gitignore

* Fix compilation errors

* Update AppConfig.cs

* Remove  unused method

* added the migrations folder and the database context with only the user as of now

* refactor: orm for UserRepository, works as intended

* changed some domain stuff, removed inheritance from Buyer, not needed, project builds without it, I removed it because EF would attempt to create a single table out of User and Buyer with a Discriminator column because of the inheritance, it also simplified stuff to keep the db logic in place

* fix: setter for id for the Entity Framework to work

* Update UserApiController.cs

- Remove unnecessary comments.
- Return the correct response type for create and update
- add ProducesResponseType for all methods
- Fix stylecop warnings.

* Add web api warnings for undeclared status codes

* Move Buyer, Seller, User proxy to the correct namespace

* Add comments for UserProxyRepository

* Fix missing using declarations

* prefix API related addresses with api

* refactor: BuyerRepository with all the needed changes across code, added some data models in order to keep the domain unchanged, converted in the repository to the actual domain objects where needed

* refactor: stylecop

* Implement BuyerRepository proxy and controller

* Use ApiBaseRoute constant

* refactor: SellerRepository with the changes needed

* Create empty SellerApiController

* Implement SellerRepository proxy and controller

* DummyProductRepo IMPORTANT MESSAGE

* refactor: ContractRenewalRepository + needed changes for stuff to work as expected, big change in productType (changed from int to string across entire application, not only Server), also made a modification to a method involving the renewal of a contract, where the byte array parameter was not used at all, so I removed it

* refactor: PDFRepository

* Remove stale comments

* Delete WeatherForecastController.cs

* Remove proxies from server

* Fix runtime errors in ContractRenewalRepository

* Implement proxy and controller for ContractRenewalRepository

* feat: added the following tables to the code first approach: TrackedOrders, OrderCheckpoints, WaitlistProducts, UserWaitList, DummyCard, DummyWallet, BuyerCartItems with the needed changes and Server entities + left a lost of comments justifying every change which could be questionable

* refactor: ContractRepository

* Implement ContractRepository proxy and controller

* Implement PDFRepository proxy and controller

* Update Program.cs

* Make internal method private

* refactor: NotificationRepository + modification to interface, 2 methods not neededanymore + modification to the repository proxy, those will not be needed anymore + got rid of those methods from the service as well, project builds just fine, removed Disposable implementatiton from INotificationContentService

* Implement NotificationRepository proxy and controller

* Implement WaitListRepository proxy and controller

* Implement OrderHistoryRepository proxy and controller

* refactor: OrderRepository + changed a signature to allow null values if the strings had the default value null in the first place (method GetOrdersWithProductInfoAsync)

* Implement OrderRepository proxy and controller

* refactor: OrderSummaryRepository + changed the documentation of methods in the interface because it is inherited in the actual class (needed to specify that it could throw an exception)

* Implement OrderSummaryRepository proxy and controller

* refactor: major refactoring, namespaces were wrong, loaded the repos in Program.cs, will uncomment the rest as I implement

* update comments

* Put all data transfer objects in a namespace

* refactor: TrackedOrderRepository + uncommented line in Server/Program.cs

* fix: cleared some warnings

* Implement TrackedOrderRepository proxy and controller

* Implement DummyWalletRepository proxy and controller

* Implement DummyCardRepository proxy and controller

* Implement DummyProductRepository proxy and controller

* Fix runtime errors

* refactor: WaitListRepository + refactoring in code and in application logic, in short, Lirili Larila came by and wispered in my ear that productId != productWaitListId

* refactor: DummyCardRepository + added index on CardNumber => added new migration, why? because Golubiro Spioniro is stealing our cards(also in the comments to se where he gets the information from)

* fix: forgot to uncomment 2 repos in Program.cs, project builds like I built 90s in Fortnite in 2020

* refactor: DummyWalletRepository + added a new foreign key to buyer table (one to one relation) to be able to implement the repo, Maria did not have this column in the beginning

* fix: added a one to many relationship between DummyCard and Buyer (a buyer can have many cards) so that the code is more mainaintable in the future

* Merge from alex.cs

* refactor: some refactoring in repository, no vars in repos I refactored, also checked that every throw is signaled in the method documentation. NOTE: The only 2 repos I did not refactor are OrderHistory and DummyProduct because this implies DummyProduct -> Product refactoring which I was not assigned to do. If help is needed for the two repos after the refactoring is done, I will gladly do it.

* Update WaitListProxyRepository.cs

* Update ContractRenewalApiController.cs

* fix: user creation for seller works, update seller profile works as expected+ deleted useless code

* fix: Buyer Profile Save work as intended in all cases, buyer login works

* Merged shoppingcart

* REFACTOR DUMMY PRODUCT: GOLUBIRO SPIONIRO SAYS DUMMY PRODUCTS ARE FOR LOSERS

* moved stuff from place to place
proxy repo & service & helper -> shared class library
+ update namespaces and imports (projects builds)

* small fix

* refactor-fix: forgot to update the namespaces of repos after moving them to SharedClassLibrary

* refactor: INotificationRepository now has async methods and all code using those methods was updated accordingly

* Fix: ShoppingCartProxyRepository calls incorrect api urls

* Fix table column names in sql query

* Fix: Product can't be used as a dictionary key

* Return error messages from the api

* Remove nonexistent namespace

* refactor-fix: forgot to update the namespaces of repos after moving them to SharedClassLibrary

* Fix compilation error

* Remove hardcoded buyerId

* fix: namespace for Interface Repo was wrong, updated it and the places it was used in

* refactor: modified the folder name in SharedClassLibrary from Repository to ProxyRepository so it is more clear throught the application what type of repository it is used, also updated all namespaces and imports throught the application, project builds

* refactor: OrderHistoryRepository

* refactor: ProductRepository

* fix: this file was in the wrong place and also duplicated(it already existed in shared class library, project builds

* fix: warning fix in ProductRepository

* refactor: ShoppingCartRepository

* refactor: updated params in lambda functions because Imre :)

---------

+++ Server/Controllers/DummyCardApiController.cs
@@ -54,9 +54,9 @@ namespace Server.Controllers
         /// <param name="cardNumber">The card number.</param>
         /// <returns>The card balance.</returns>
         [HttpGet("{cardNumber}/balance")]
-        [ProducesResponseType(typeof(float), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(double), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<float>> GetCardBalance(string cardNumber)
+        public async Task<ActionResult<double>> GetCardBalance(string cardNumber)
         {
             try
             {
@@ -78,7 +78,7 @@ namespace Server.Controllers
         [HttpPut("{cardNumber}/balance")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> UpdateCardBalance(string cardNumber, [FromBody] float balance)
+        public async Task<IActionResult> UpdateCardBalance(string cardNumber, [FromBody] double balance)
         {
             try
             {
