+++ Server/Repository/OrderSummaryRepository.cs
@@ -26,7 +26,7 @@ namespace Server.Repository
         }
 
         /// <inheritdoc/>
-        public async Task UpdateOrderSummaryAsync(int id, float subtotal, float warrantyTax, float deliveryFee, float finalTotal, string fullName, string email, string phoneNumber, string address, string postalCode, string additionalInfo, string contractDetails)
+        public async Task UpdateOrderSummaryAsync(int id, double subtotal, double warrantyTax, double deliveryFee, double finalTotal, string fullName, string email, string phoneNumber, string address, string postalCode, string additionalInfo, string contractDetails)
         {
             OrderSummary? orderSummary = await this.dbContext.OrderSummary.FindAsync(id)
                                                 ?? throw new KeyNotFoundException($"UpdateOrderSummaryAsync: OrderSummary with ID {id} not found");
