+++ SharedClassLibrary/IRepository/IDummyCardRepository.cs
@@ -5,7 +5,7 @@ namespace SharedClassLibrary.IRepository
     public interface IDummyCardRepository
     {
         Task DeleteCardAsync(string cardNumber);
-        Task<float> GetCardBalanceAsync(string cardNumber);
-        Task UpdateCardBalanceAsync(string cardNumber, float balance);
+        Task<double> GetCardBalanceAsync(string cardNumber);
+        Task UpdateCardBalanceAsync(string cardNumber, double balance);
     }
 }
